requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
# Optional backbone zoo (if available)
timm>=0.9.8
# Data & metrics
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
# Plotting (optional for confusion matrix, curves)
matplotlib>=3.8.0
# For VLM workflows (install only when running LLaVA-Med)
transformers>=4.42.0
accelerate>=0.33.0
peft>=0.12.0
bitsandbytes>=0.43.0
